Use latest upstream commit for validator
This had to be changed, because the GH Action otherwise has it hard
coded that it should go to the commit/tag/branch SHA in
Materials-Consortia/optimade-python-tools, however, this branch is under
CasperWA/optimade-python-tools, and hence the given SHA can never be
found if it's different/new.
The solution is to instead use the latest commit from
Materials-Consortia/optimade-python-tools, since we are not changing the
validator in this branch in any case.